Give up on creating volume dirs beforehand because of postgres
Postgres' initdb won't accept a dir with a dotfile. To recap, the only
way for docker to mount a volume dir as a container runtime user, is for
it to pre-exist with those permissions. We will need to wrap the usage
of the compose file with a script to create the volume dirs.